hawkbit-client: enable TCP keep-alive probing for get_binary()
Keep-alive probes allow us to determine whether a connection is still up
and running or not. Depending on the outage, this enables us to fail
faster.

This is especially useful when CURLOPT_LOW_SPEED_LIMIT and
CURLOPT_LOW_SPEED_TIME will be configurable by the user in a future
commit: if the configuration allows download rates over long timespans,
there will be no intervention for hanging connections whatsoever (until
the configured timespan is exceeded).

Signed-off-by: Bastian Krause <bst@pengutronix.de>