test: run cancel/install tests also in streaming mode
Most of the install and cancel test can be run in "normal" download mode
and streaming installation mode. Use pytest.mark.parametrize to realize
that. Use the 'download'/'streaming' strings as mode argument. That way
the tests get a human-readable parameter, e.g.:

  test/test_install.py::test_install_maintenance_window[download]
  test/test_install.py::test_install_maintenance_window[streaming]

While at it, disable the anonymous download option explicitly. That way
it is ensured that hawkBit actually checks the authentication header on
bundle download. This is the default, but better set it explicitly.

Signed-off-by: Bastian Krause <bst@pengutronix.de>