github/README: set hawkbit security options for RAUC HTTP streaming
RAUC HTTP streaming works by translating NBD read requests into HTTP
range requests [1]. The default squashfs block size is 128 KB [2]. The
ranges requested happen to be in the same order of magnitude.

With this in mind, certain default configurations in hawkBit do not work
out of the box:

hawkBit tries to prevent DoS attacks [3]:

  "hawkBit owns the feature of guarding itself from DoS attacks, a DoS
  filter. It reduces the maximum number of requests per seconds which can
  be configured for read and write requests." [4]

This might affect RAUC HTTP bundle streaming because it causes a high
number of HTTP range requests within a time period:

  Registered DOS attack! Client [0:0:0:0:0:0:0:1] is above configured READ request threshold (200)!

The requests are authenticated, so it should be safe to disable the DoS
protection.

hawkBit generats an "ActionStatus" entity for each HTTP range request.
When using RAUC HTTP streaming, a huge number of those requests are
performed. It is likely that the default quota limit of allowed
"ActionStatus" entities is hit:

  Cannot assign 1 ActionStatus entities to Action '1' because of the configured quota limit 1000. Currently, there are 1000 ActionStatus entities assigned.

To allow requesting an unlimited number of ranges per bundle, disable
the "status entries per action quota" altogether.

Generating an "ActionStatus" per range request is probably not a good
idea for this use case. This needs discussion with upstream.

[1] https://rauc.readthedocs.io/en/latest/advanced.html#http-streaming
[2] https://www.kernel.org/doc/html/latest/filesystems/squashfs.html
[3] https://github.com/eclipse/hawkbit/blob/master/hawkbit-http-security/src/main/java/org/eclipse/hawkbit/security/DosFilter.java
[4] https://www.eclipse.org/hawkbit/guides/clustering/

Signed-off-by: Bastian Krause <bst@pengutronix.de>